refactor: Move checklists from data/ to content/
Moves all checklist JSON files and index.ts from data/checklists/ to
content/checklists/ to be consistent with the rest of the content on
the site (posts, guides, exercises, quizzes, news, etc.).

Updated import paths in:
- app/checklists/page.tsx
- app/checklists/[slug]/page.tsx
- scripts/generate-search-index.ts
- scripts/generate-checklist-images-svg.ts

Closes #747
